teste azure schedule

+++ src/main/java/br/com/demo/demo/infrastructure/DomainExemploRepository.java
@@ -3,5 +3,5 @@ package br.com.demo.demo.infrastructure;
 import br.com.demo.demo.domain.DomainExemplo;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface DomainExemploRepository extends PagingAndSortingRepository<DomainExemplo,Integer> {
+public interface DomainExemploRepository extends PagingAndSortingRepository<DomainExemplo,Long> {
 }
